refactor: Schedule 서비스 분리
