+++ src/main/java/com/example/team1_be/domain/Schedule/Service/ScheduleService.java
@@ -19,7 +19,6 @@ import com.example.team1_be.domain.Apply.ApplyStatus;
 import com.example.team1_be.domain.Day.Day;
 import com.example.team1_be.domain.Day.DayService;
 import com.example.team1_be.domain.Group.Group;
-import com.example.team1_be.domain.Group.Service.GroupService;
 import com.example.team1_be.domain.Schedule.DTO.FixSchedule;
 import com.example.team1_be.domain.Schedule.DTO.GetDailyFixedApplies;
 import com.example.team1_be.domain.Schedule.DTO.GetFixedWeeklySchedule;
@@ -34,7 +33,6 @@ import com.example.team1_be.domain.Schedule.Recommend.WeeklySchedule.Recommended
 import com.example.team1_be.domain.Schedule.Recommend.WorktimeApply.RecommendedWorktimeApply;
 import com.example.team1_be.domain.Schedule.Recommend.WorktimeApply.RecommendedWorktimeApplyService;
 import com.example.team1_be.domain.Schedule.Schedule;
-import com.example.team1_be.domain.Schedule.ScheduleRepository;
 import com.example.team1_be.domain.User.User;
 import com.example.team1_be.domain.User.UserService;
 import com.example.team1_be.domain.Week.Week;
@@ -54,14 +52,15 @@ public class ScheduleService {
 	private final int NUM_DAYS_OF_WEEK = 7;
 
 	private final UserService userService;
-	private final GroupService groupService;
-	private final ScheduleRepository scheduleRepository;
 	private final WeekService weekService;
 	private final DayService dayService;
 	private final WorktimeService worktimeService;
 	private final ApplyService applyService;
 	private final RecommendedWorktimeApplyService recommendedWorktimeApplyService;
 	private final RecommendedWeeklyScheduleService recommendedWeeklyScheduleService;
+
+	private final ScheduleReadOnlyService scheduleReadOnlyService;
+	private final ScheduleWriteService scheduleWriteService;
 
 	@Transactional
 	public void recruitSchedule(User user, RecruitSchedule.Request request) {
@@ -74,7 +73,7 @@ public class ScheduleService {
 		Schedule schedule = Schedule.builder()
 			.group(group)
 			.build();
-		scheduleRepository.save(schedule);
+		scheduleWriteService.createSchedule(schedule);
 
 		Week week = weekService.createWeek(schedule, request.getWeekStartDate());
 
@@ -105,7 +104,7 @@ public class ScheduleService {
 	public WeeklyScheduleCheck.Response weeklyScheduleCheck(User user, LocalDate request) {
 		Group group = userService.findGroupByUser(user);
 
-		Schedule schedule = findByGroup(group);
+		Schedule schedule = scheduleReadOnlyService.findByGroup(group);
 
 		Week week = user.getIsAdmin() ?
 			weekService.findByScheduleIdStartDateAndStatus(schedule, request, WeekRecruitmentStatus.STARTED) :
@@ -125,7 +124,7 @@ public class ScheduleService {
 
 	public GetFixedWeeklySchedule.Response getFixedWeeklySchedule(User user, YearMonth requestMonth, Long memberId) {
 		User member = userService.findById(memberId);
-		Schedule schedule = findByGroup(member.getGroup());
+		Schedule schedule = scheduleReadOnlyService.findByGroup(member.getGroup());
 
 		LocalDate date = LocalDate.of(requestMonth.getYear(), requestMonth.getMonth(), 1);
 		LocalDate toDate = LocalDate.of(requestMonth.getYear(), requestMonth.getMonth(), 1).plusMonths(1);
@@ -145,7 +144,7 @@ public class ScheduleService {
 	public RecommendSchedule.Response recommendSchedule(User user, LocalDate date) {
 		Group group = userService.findGroupByUser(user);
 
-		Schedule schedule = findByGroup(group);
+		Schedule schedule = scheduleReadOnlyService.findByGroup(group);
 
 		List<Worktime> weeklyWorktimes = worktimeService.findByStartDateAndSchedule(date, schedule);
 
@@ -208,7 +207,7 @@ public class ScheduleService {
 
 	public GetDailyFixedApplies.Response getDailyFixedApplies(User user, LocalDate selectedDate) {
 		Group group = userService.findGroupByUser(user);
-		Schedule schedule = findByGroup(group);
+		Schedule schedule = scheduleReadOnlyService.findByGroup(group);
 
 		LocalDate date = selectedDate.minusDays(selectedDate.getDayOfWeek().getValue() - 1);
 		int dayOfWeek = selectedDate.getDayOfWeek().getValue();
@@ -227,7 +226,7 @@ public class ScheduleService {
 
 	public GetFixedWeeklySchedule.Response getUsersFixedWeeklySchedule(User user, YearMonth requestMonth) {
 		Group group = userService.findGroupByUser(user);
-		Schedule schedule = findByGroup(group);
+		Schedule schedule = scheduleReadOnlyService.findByGroup(group);
 
 		LocalDate date = LocalDate.of(requestMonth.getYear(), requestMonth.getMonth(), 1);
 		LocalDate toDate = LocalDate.of(requestMonth.getYear(), requestMonth.getMonth(), 1).plusMonths(1);
@@ -246,7 +245,7 @@ public class ScheduleService {
 	public LoadLatestSchedule.Response loadLatestSchedule(User user, LocalDate startWeekDate) {
 		Group group = userService.findGroupByUser(user);
 
-		Schedule schedule = findByGroup(group);
+		Schedule schedule = scheduleReadOnlyService.findByGroup(group);
 
 		List<Week> latestWeeks = weekService.findLatestByScheduleAndStatus(schedule, WeekRecruitmentStatus.ENDED);
 		if (latestWeeks.isEmpty()) {
@@ -261,7 +260,7 @@ public class ScheduleService {
 	public GetWeekStatus.Response getWeekStatus(User user, LocalDate startWeekDate) {
 		Group group = userService.findGroupByUser(user);
 
-		Schedule schedule = findByGroup(group);
+		Schedule schedule = scheduleReadOnlyService.findByGroup(group);
 		Week week = weekService.findByScheduleAndStartDate(schedule, startWeekDate);
 
 		if (week == null) {
@@ -269,10 +268,5 @@ public class ScheduleService {
 		} else {
 			return new GetWeekStatus.Response(week.getStatus());
 		}
-	}
-
-	public Schedule findByGroup(Group group) {
-		return scheduleRepository.findByGroup(group)
-			.orElseThrow(() -> new NotFoundException("스케줄을 찾을 수 없습니다."));
 	}
 }
